Refactor server menu and input handling, add modular headers
Introduced new input.h and menu.h headers and corresponding input.c and menu.c source files to modularize user input and menu logic on the server. Refactored server/main.c to use the new menu system, updated CMakeLists.txt, and removed legacy code from send_commands.c (now menu.c). Added server_set_device_state_and_update_flash for unified device state management and flash persistence. Improved code organization and maintainability.
