+++ common/functions.h
@@ -5,6 +5,7 @@
 #include "pico/stdio.h"
 #include "hardware/uart.h"
 #include "hardware/gpio.h"
+
 #include "types.h"
 
 int pico_led_init(void);
