+++ common/types.c
@@ -1,7 +1,6 @@
 #include "types.h"
 #include "config.h"
 
-// make a .json/.ini/CSV file instead of hardcoding these variables
 const uart_pin_pair_t pin_pairs_uart0[PIN_PAIRS_UART0_LEN] = {
     {0, 1},
     {12, 13},
