+++ include/server.h
@@ -2,7 +2,9 @@
 #define SERVER_H
 
 #include <stdio.h>
+
 #include "hardware/uart.h"
+
 #include "config.h"
 #include "types.h"
 
@@ -16,7 +18,6 @@ extern uint8_t active_server_connections_number;
 
 bool server_find_connections(void);
 bool server_uart_read(uart_inst_t*, uint32_t);
-void server_listen_for_commands(void);
 void server_load_running_states_to_active_clients(void);
 void server_print_running_client_state(client_t *);
 void server_print_client_preset_configuration(client_t *, uint8_t);
@@ -31,6 +32,17 @@ inline void server_print_gpio_state(uint8_t gpio_index, const client_state_t *cl
         printf("Device[%u]: gpio_number:%u is_on:%u\n", gpio_index + 1, client_state->devices[gpio_index].gpio_number, client_state->devices[gpio_index].is_on);
     }
 }
+
+/**
+ * @brief Sends device state to client and persists it in Flash.
+ * 
+ * @param pin_pair         TX/RX GPIO pair for the UART connection.
+ * @param uart_instance    UART instance used for the connection.
+ * @param gpio_number      GPIO number on the client to control.
+ * @param device_state     Desired state (true = ON, false = OFF).
+ * @param flash_index      Index in flash memory to store the state.
+ */
+void server_set_device_state_and_update_flash(uart_pin_pair_t , uart_inst_t*, uint8_t, bool, uint32_t);
 
 #define SERVER_SECTOR_SIZE    4096
 #define SERVER_PAGE_SIZE      256
